+++ contracts/protocol/libraries/types/PositionManagerStorage.sol
@@ -10,8 +10,8 @@ contract PositionManagerStorage {
     using TickPosition for TickPosition.Data;
     using LiquidityBitmap for mapping(uint128 => uint256);
 
-    uint256 public basisPoint; //0.01
-    uint256 public BASE_BASIC_POINT;
+    uint64 public basisPoint; //0.01
+    uint64 public BASE_BASIC_POINT;
     // fee = quoteAssetAmount / tollRatio (means if fee = 0.001% then tollRatio = 100000)
     uint256 tollRatio;
 
@@ -52,8 +52,8 @@ contract PositionManagerStorage {
 
     struct ReserveSnapshot {
         uint128 pip;
-        uint256 timestamp;
-        uint256 blockNumber;
+        uint64 timestamp;
+        uint64 blockNumber;
     }
 
     enum TwapCalcOption {
